+++ source/core/FtpFileSystem.h
@@ -143,8 +143,7 @@ protected:
     const TFtpsCertificateData & Data, int & RequestResult);
   bool __fastcall HandleAsynchRequestNeedPass(
     struct TNeedPassRequestData & Data, int & RequestResult);
-  bool __fastcall HandleListData(const wchar_t * Path, const TListDataEntry * Entries,
-    unsigned int Count);
+  bool __fastcall HandleListData(const wchar_t * Path, const TListDataEntry * Entries, size_t Count);
   bool __fastcall HandleTransferStatus(bool Valid, __int64 TransferSize,
     __int64 Bytes, bool FileTransfer);
   bool __fastcall HandleReply(int Command, unsigned int Reply);
